# packages/client/openapi.json
{
  "components": {
    "schemas": {
      "Alternative": {
        "properties": {
          "reason": {
            "description": "How this table differs from the selection.",
            "title": "Reason",
            "type": "string"
          },
          "table_id": {
            "description": "ACS table ID.",
            "title": "Table Id",
            "type": "string"
          },
          "title": {
            "description": "Published table title.",
            "title": "Title",
            "type": "string"
          },
          "universe": {
            "description": "Published universe of this table.",
            "title": "Universe",
            "type": "string"
          }
        },
        "required": [
          "table_id",
          "title",
          "universe",
          "reason"
        ],
        "title": "Alternative",
        "type": "object"
      },
      "AskRequest": {
        "properties": {
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultPlan"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional typed override of table, years, geography, and overlapping ACS5."
          },
          "question": {
            "description": "The user's question. Leading and trailing whitespace is stripped.",
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "properties": {
          "alternatives": {
            "description": "Related tables, with why they differ.",
            "items": {
              "$ref": "#/components/schemas/Alternative"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "answer": {
            "description": "Natural-language answer.",
            "title": "Answer",
            "type": "string"
          },
          "attempted_years": {
            "description": "Vintages a Census request was issued for.",
            "items": {
              "type": "integer"
            },
            "title": "Attempted Years",
            "type": "array"
          },
          "chart": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChartSpec"
              },
              {
                "type": "null"
              }
            ],
            "description": "Validated chart roles, or null."
          },
          "chart_unavailable": {
            "default": false,
            "description": "Model chart failed validation.",
            "title": "Chart Unavailable",
            "type": "boolean"
          },
          "comparisons": {
            "description": "Paired estimates with MOE_diff.",
            "items": {
              "$ref": "#/components/schemas/Comparison"
            },
            "title": "Comparisons",
            "type": "array"
          },
          "failed_years": {
            "description": "Attempted vintages that failed or timed out.",
            "items": {
              "type": "integer"
            },
            "title": "Failed Years",
            "type": "array"
          },
          "geoid": {
            "description": "AFFGEOID of the selected geography; empty if many.",
            "title": "Geoid",
            "type": "string"
          },
          "legs": {
            "description": "Per-request outcome in requested order.",
            "items": {
              "$ref": "#/components/schemas/RequestLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "moe": {
            "description": "Per-row 90% margins keyed to each M.",
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Moe",
            "type": "array"
          },
          "omission_reasons": {
            "description": "Reason code per omitted year.",
            "items": {
              "type": "string"
            },
            "title": "Omission Reasons",
            "type": "array"
          },
          "omitted_years": {
            "description": "Requested years that were not attempted.",
            "items": {
              "type": "integer"
            },
            "title": "Omitted Years",
            "type": "array"
          },
          "plan": {
            "$ref": "#/components/schemas/ResultPlan",
            "description": "Executed table, years, and geographies for this answer."
          },
          "requested_years": {
            "description": "Years asked of fetch_data.",
            "items": {
              "type": "integer"
            },
            "title": "Requested Years",
            "type": "array"
          },
          "rows": {
            "description": "Census rows; each carries GEO_ID.",
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "succeeded_years": {
            "description": "Attempted vintages that returned rows.",
            "items": {
              "type": "integer"
            },
            "title": "Succeeded Years",
            "type": "array"
          },
          "table_id": {
            "description": "Selected ACS table ID.",
            "title": "Table Id",
            "type": "string"
          },
          "universe": {
            "description": "Published universe of the selected table.",
            "title": "Universe",
            "type": "string"
          },
          "urls": {
            "description": "Key-redacted Census API URLs, one per attempted request.",
            "items": {
              "type": "string"
            },
            "title": "Urls",
            "type": "array"
          },
          "warnings": {
            "description": "Non-blocking guards (docs/requirements.md).",
            "items": {
              "$ref": "#/components/schemas/AskWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "urls",
          "requested_years",
          "attempted_years",
          "succeeded_years",
          "failed_years",
          "omitted_years",
          "omission_reasons",
          "legs",
          "rows",
          "moe",
          "geoid",
          "universe",
          "table_id",
          "alternatives",
          "comparisons",
          "warnings",
          "plan"
        ],
        "title": "AskResponse",
        "type": "object"
      },
      "AskWarning": {
        "properties": {
          "candidates": {
            "description": "Ranked executable geographies when code is ambiguous_place; else empty.",
            "items": {
              "$ref": "#/components/schemas/GeoSpec"
            },
            "title": "Candidates",
            "type": "array"
          },
          "code": {
            "description": "Machine-readable warning code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "What was raised, in one sentence.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "AskWarning",
        "type": "object"
      },
      "ChartSpec": {
        "additionalProperties": false,
        "description": "Roles over rows/moe. No duplicated data, Vega, or markup.",
        "properties": {
          "series_by": {
            "anyOf": [
              {
                "enum": [
                  "geography",
                  "variable"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Split by GEO_ID or estimate variable.",
            "title": "Series By"
          },
          "show_moe": {
            "const": true,
            "default": true,
            "description": "MOE display cannot be opted out.",
            "title": "Show Moe",
            "type": "boolean"
          },
          "title": {
            "description": "Plain-text title.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "line for years; bar for geographies.",
            "enum": [
              "line",
              "bar"
            ],
            "title": "Type",
            "type": "string"
          },
          "x": {
            "description": "Axis from row year or GEO_ID.",
            "enum": [
              "year",
              "geography"
            ],
            "title": "X",
            "type": "string"
          },
          "y": {
            "const": "estimate",
            "description": "Estimate columns; never copied values.",
            "title": "Y",
            "type": "string"
          }
        },
        "required": [
          "type",
          "x",
          "y",
          "title"
        ],
        "title": "ChartSpec",
        "type": "object"
      },
      "Comparison": {
        "properties": {
          "distinguishable": {
            "description": "True when abs(estimate_a - estimate_b) exceeds threshold.",
            "title": "Distinguishable",
            "type": "boolean"
          },
          "estimate_a": {
            "description": "Published estimate for geoid_a.",
            "title": "Estimate A",
            "type": "string"
          },
          "estimate_b": {
            "description": "Published estimate for geoid_b.",
            "title": "Estimate B",
            "type": "string"
          },
          "geoid_a": {
            "description": "AFFGEOID of the first leg.",
            "title": "Geoid A",
            "type": "string"
          },
          "geoid_b": {
            "description": "AFFGEOID of the second leg.",
            "title": "Geoid B",
            "type": "string"
          },
          "moe_a": {
            "description": "90% margin for estimate_a.",
            "title": "Moe A",
            "type": "string"
          },
          "moe_b": {
            "description": "90% margin for estimate_b.",
            "title": "Moe B",
            "type": "string"
          },
          "shared_sample": {
            "description": "True when the compared geographies nest and share ACS sample.",
            "title": "Shared Sample",
            "type": "boolean"
          },
          "threshold": {
            "description": "MOE_diff = sqrt(moe_a^2 + moe_b^2).",
            "title": "Threshold",
            "type": "string"
          },
          "variable": {
            "description": "Estimate variable compared.",
            "title": "Variable",
            "type": "string"
          },
          "year": {
            "description": "Vintage end year when both legs share one; else empty.",
            "title": "Year",
            "type": "string"
          }
        },
        "required": [
          "variable",
          "geoid_a",
          "geoid_b",
          "year",
          "estimate_a",
          "estimate_b",
          "moe_a",
          "moe_b",
          "threshold",
          "distinguishable",
          "shared_sample"
        ],
        "title": "Comparison",
        "type": "object"
      },
      "Conversation": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "thread_id": {
            "format": "uuid",
            "title": "Thread Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/Turn"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "thread_id",
          "title",
          "created_at",
          "expires_at",
          "turns"
        ],
        "title": "Conversation",
        "type": "object"
      },
      "ConversationInfo": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "thread_id": {
            "format": "uuid",
            "title": "Thread Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "title",
          "created_at",
          "expires_at"
        ],
        "title": "ConversationInfo",
        "type": "object"
      },
      "GeoSpec": {
        "description": "Executable geography. `for`/`in` come from metadata, not from prose.",
        "properties": {
          "codes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Codes",
            "type": "object"
          },
          "dataset": {
            "default": "acs5",
            "title": "Dataset",
            "type": "string"
          },
          "for_spec": {
            "default": "",
            "title": "For Spec",
            "type": "string"
          },
          "geoid": {
            "default": "",
            "title": "Geoid",
            "type": "string"
          },
          "in_spec": {
            "default": "",
            "title": "In Spec",
            "type": "string"
          },
          "level": {
            "default": "",
            "title": "Level",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "vintage": {
            "default": 0,
            "title": "Vintage",
            "type": "integer"
          }
        },
        "title": "GeoSpec",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "RequestLeg": {
        "properties": {
          "detail": {
            "description": "Redacted reason when ok is false; empty on success.",
            "title": "Detail",
            "type": "string"
          },
          "for_spec": {
            "default": "",
            "description": "Census for= clause for this geography leg; empty when the built URL is used.",
            "title": "For Spec",
            "type": "string"
          },
          "ok": {
            "description": "Whether this HTTP call returned parseable rows.",
            "title": "Ok",
            "type": "boolean"
          },
          "status_code": {
            "description": "Census HTTP status; 0 on timeout or transport failure.",
            "title": "Status Code",
            "type": "integer"
          },
          "url": {
            "description": "Key-redacted Census API URL for this year.",
            "title": "Url",
            "type": "string"
          },
          "year": {
            "description": "Vintage end year for this Census request.",
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "url",
          "ok",
          "status_code",
          "detail"
        ],
        "title": "RequestLeg",
        "type": "object"
      },
      "ResultPlan": {
        "description": "Executed table, variables, years, and geographies. Not extracted from prose.",
        "properties": {
          "allow_overlapping_acs5": {
            "default": false,
            "description": "Consecutive ACS5 explicitly allowed.",
            "title": "Allow Overlapping Acs5",
            "type": "boolean"
          },
          "dataset": {
            "default": "acs5",
            "description": "acs5 or acs1 actually requested.",
            "title": "Dataset",
            "type": "string"
          },
          "geographies": {
            "description": "Ordered resolved geos.",
            "items": {
              "$ref": "#/components/schemas/GeoSpec"
            },
            "title": "Geographies",
            "type": "array"
          },
          "requested_years": {
            "description": "Fetch request years.",
            "items": {
              "type": "integer"
            },
            "title": "Requested Years",
            "type": "array"
          },
          "table_id": {
            "default": "",
            "description": "Selected ACS table ID.",
            "title": "Table Id",
            "type": "string"
          },
          "variables": {
            "description": "Estimate IDs (E) from the URL.",
            "items": {
              "type": "string"
            },
            "title": "Variables",
            "type": "array"
          },
          "years": {
            "description": "Vintages a request was issued for.",
            "items": {
              "type": "integer"
            },
            "title": "Years",
            "type": "array"
          }
        },
        "title": "ResultPlan",
        "type": "object"
      },
      "Turn": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/AskResponse"
          }
        },
        "required": [
          "question",
          "plan",
          "response",
          "created_at"
        ],
        "title": "Turn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "census-concierge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/ask": {
      "post": {
        "operationId": "ask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Ask"
      }
    },
    "/conversations": {
      "post": {
        "operationId": "create_conversation",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-User-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "persistence unavailable"
          }
        },
        "summary": "Post Conversation"
      }
    },
    "/conversations/{thread_id}": {
      "get": {
        "operationId": "get_conversation",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-User-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "conversation not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "persistence unavailable"
          }
        },
        "summary": "Get Conversation"
      }
    },
    "/conversations/{thread_id}/turns": {
      "post": {
        "operationId": "append_turn",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-User-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "conversation not found"
          },
          "409": {
            "description": "conversation full"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "persistence unavailable"
          }
        },
        "summary": "Post Turn"
      }
    }
  }
}